Fix 64bit compile.

+++ src/pf/chunks/CINP/SceneData.cpp
@@ -537,7 +537,7 @@ namespace gw2f {
 			SpeciesResourceDataV36::SpeciesResourceDataV36( ) {
 			}
 
-			SpeciesResourceDataV36::SpeciesResourceDataV36( const byte* p_data, uint32 p_size, const byte** po_pointer ) {
+			SpeciesResourceDataV36::SpeciesResourceDataV36( const byte* p_data, size_t p_size, const byte** po_pointer ) {
 				auto pointer = assign( p_data, p_size );
 				if ( po_pointer ) { *po_pointer = pointer; }
 			}
@@ -557,7 +557,7 @@ namespace gw2f {
 				return *this;
 			}
 
-			const byte* SpeciesResourceDataV36::assign( const byte* p_data, uint32 p_size ) {
+			const byte* SpeciesResourceDataV36::assign( const byte* p_data, size_t p_size ) {
 				p_data = helpers::read( p_data, p_size, speciesId );
 				p_data = helpers::read( p_data, p_size, name );
 				p_data = helpers::read( p_data, p_size, modelId );
